requirements.txt
discord.py
psutil
googletrans
